+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - Improve error message
 - Update to dart 3.8.0
 - Don't add log type to commit message
-- Some change 
+- Some change
+- When not existing gg do push creates an upstream branch on the remote
 
 ### Fixed